feat: move GetAppDataDir to platform specif path (+small WIN32 guards)
