+++ RuneHelper/CMakeLists.txt
@@ -118,6 +118,7 @@ add_executable(runehelper
   ocr/OCR.cpp
   price/PriceCache.cpp
   ui/Overlay.cpp
+  platform/linux/PlatformPaths.cpp
   platform/linux/UIManager.cpp
   platform/linux/ScreenCapture.cpp
   platform/linux/RegionSelect.cpp
