+++ RuneHelper/RuneHelper.cpp
@@ -1,10 +1,16 @@
+#ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
 #define NOMINMAX
 #include <windows.h>
+#endif
 
 #include "core/RuneHelperApp.h"
 
+#ifdef _WIN32
 int WINAPI WinMain(HINSTANCE, HINSTANCE, LPSTR, int)
+#else
+int main()
+#endif
 {
     RuneHelperApp app;
     return app.Run();
