+++ RuneHelper/core/ConfigManager.cpp
@@ -1,8 +1,5 @@
 #include "ConfigManager.h"
 
-#include <windows.h>
-#include <shlobj.h>
-
 #include <filesystem>
 #include <fstream>
