+++ RuneHelper/core/Logger.cpp
@@ -2,8 +2,11 @@
 
 #define WIN32_LEAN_AND_MEAN
 #define NOMINMAX
+#ifdef _WIN32
 #include <windows.h>
-#include <shlobj.h>
+#endif
+
+#include "platform/PlatformPaths.h"
 
 #include <chrono>
 #include <filesystem>
@@ -31,16 +34,7 @@ Logger::~Logger()
 
 bool Logger::Init()
 {
-    PWSTR appData = nullptr;
-
-    if (FAILED(SHGetKnownFolderPath(FOLDERID_RoamingAppData, 0, nullptr, &appData)))
-        return false;
-
-    std::filesystem::path dir(appData);
-    CoTaskMemFree(appData);
-
-    dir /= "Denz";
-    dir /= "RuneHelper";
+    std::filesystem::path dir = GetAppDataDir();
 
     std::filesystem::create_directories(dir);
 
@@ -105,7 +99,11 @@ std::string Logger::TimeNow()
     auto t = std::chrono::system_clock::to_time_t(now);
 
     std::tm tm{};
+#ifdef _WIN32
     localtime_s(&tm, &t);
+#else
+    localtime_r(&t, &tm);
+#endif
 
     std::ostringstream ss;
