+++ RuneHelper/platform/PlatformPaths.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <filesystem>
+
+std::filesystem::path GetAppDataDir();
